requirements.txt
streamlit==1.28.1
easyocr==1.7.0
opencv-python==4.8.1.78
pandas==2.1.1
Pillow==10.0.1
numpy==1.24.3
